requirements.txt
beautifulsoup4==4.10.0
numpy==1.21.5
onnxruntime==1.11.0
pandas==1.3.5
pillow==8.4.0
python-dateutil==2.8.2
requests==2.26.0
dataparser==23.2.1